responsive page product detail
