+++ src/Assets/Styles/page/ProductDetail/_productDetail.scss
@@ -10,7 +10,7 @@
     margin-top: $header-tablet-height;
   }
   @include mobile {
-      margin-top: $header-mobile-height;
+    margin-top: $header-mobile-height;
   }
   .title_banner_product {
     text-align: center;
@@ -40,12 +40,8 @@
 .layout {
   padding: 80px 80px;
   background: none;
-  .image_product_detail_gallery {
-    //   position: fixed;
-    top: 94px;
-    left: 80px;
-    width: 374px;
-    transform: matrix(1, 0, 0, 1, 0, 0);
+  .layout_product_detail {
+    padding: "0 22px 0 0" !important;
   }
 
   .product_detail {
@@ -185,36 +181,36 @@
     width: 100%;
     vertical-align: middle;
     margin: 65px 0 0;
-  
+
     // customs ant design ( Tabs )
 
-    .ant-tabs-nav-list{
+    .ant-tabs-nav-list {
       margin-bottom: 10px;
     }
 
-    .ant-tabs-nav{
+    .ant-tabs-nav {
       margin-bottom: 0 !important;
     }
 
-    .description_tab{
-      .content_tab_description{
-        p{
+    .description_tab {
+      .content_tab_description {
+        p {
           text-align: center;
           margin: 28px auto 0;
           font-size: $txt-p;
         }
       }
-      .content_tab_infomation{
+      .content_tab_infomation {
         margin: 28px auto 0;
         width: auto;
         min-width: 630px;
-        p{
+        p {
           margin: 0;
           font-size: $txt-p;
         }
       }
 
-      .content_tab_reviews{
+      .content_tab_reviews {
         position: relative;
         display: inline-block;
         vertical-align: middle;
@@ -222,11 +218,11 @@
         width: auto;
         min-width: 630px;
         max-width: 880px;
-        p{
+        p {
           text-align: center;
           font-size: $txt-p;
         }
-        .rely_title{
+        .rely_title {
           display: flex;
           justify-content: space-between;
           font-size: 25px;
@@ -236,7 +232,5 @@
         }
       }
     }
-
   }
 }
-
