+++ src/Page/Site/Page/ProductDetail/ProductDetail.jsx
@@ -6,55 +6,18 @@ import {
     Col,
     Image,
     Rate,
-    Comment,
     Tabs,
-    Avatar,
-    List,
 } from 'antd';
 import { BannerProduct } from '../../Components/Common/Banner/banner';
 import FormReview from './Components/FormReview';
 import RelatedProducts from './Components/RelatedProducts';
+import Comments from './Components/Comments';
 
 
 export default function ProductDetail() {
 
     const [visible, setVisible] = useState(false);
-    const data = [
-        {
-            author: 'Michael Lord - New York Times',
-            rate: 4,
-            content: (
-                <p>
-                    "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea."
-                </p>
-            ),
-            imageBottomComment: 'https://chapterone.qodeinteractive.com/wp-content/uploads/2019/07/review-1.png',
-        },
-        {
-            author: 'Joanne Smith - Huffington Post',
-            rate: 3,
-            content: (
-                <p>
-                    We supply a series of design principles, practical patterns and high quality design
-                    resources (Sketch and Axure), to help people create their product prototypes beautifully and
-                    efficiently.
-                </p>
-            ),
-            imageBottomComment: 'https://chapterone.qodeinteractive.com/wp-content/uploads/2019/07/review-2.png',
-        },
-        {
-            author: 'Minnie Loyd - The Guardian',
-            rate: 5,
-            content: (
-                <p>
-                    We supply a series of design principles, practical patterns and high quality design
-                    resources (Sketch and Axure), to help people create their product prototypes beautifully and
-                    efficiently.
-                </p>
-            ),
-            imageBottomComment: 'https://chapterone.qodeinteractive.com/wp-content/uploads/2019/07/review-3.png',
-        },
-    ];
+
     const { TabPane } = Tabs;
 
     return (
@@ -65,8 +28,18 @@ export default function ProductDetail() {
             </BannerProduct>
 
             <Layout className="layout" >
-                <Row>
-                    <Col span={8} style={{ padding: '0 22px 0 0' }}>
+                <Row >
+                    <Col
+                        xs={24}
+                        sm={24}
+                        md={24}
+                        lg={8}
+                        xl={8}
+                        // span={8}
+                        
+                        className="layout_product_detail"
+                    >
+                        {/* <Col span={8} style={{ padding: '0 22px 0 0' }}> */}
                         <div className="image_product_detail_gallery">
                             <Image
                                 preview={{ visible: false }}
@@ -82,7 +55,15 @@ export default function ProductDetail() {
 
                         </div>
                     </Col>
-                    <Col span={16} style={{ padding: '0 0 0 22px' }}>
+                    <Col
+                        xs={24}
+                        sm={24}
+                        md={24}
+                        lg={16}
+                        xl={16}
+                        // span={16}
+                        style={{ padding: '0 0 0 22px' }}
+                    >
                         <div className="product_detail">
                             <div className="title_author">
                                 <Link to="/author">by james hoffman</Link>
@@ -118,52 +99,7 @@ export default function ProductDetail() {
                             </div>
 
                         </div>
-                        <Comment
-                            className="comment_product_detail"
-                            author={<a className="author_comment">BELGIUM</a>}
-                            avatar={
-                                <Avatar
-                                    className="avatar_comment"
-                                    src="https://chapterone.qodeinteractive.com/wp-content/uploads/2019/07/avatar-2-100x100.png"
-                                    alt="Han Solo"
-                                    size={{
-                                        xs: 24,
-                                        sm: 32,
-                                        md: 40,
-                                        lg: 64,
-                                        xl: 80,
-                                        xxl: 100,
-                                    }}
-                                />
-                            }
-                            content={
-                                <div>
-                                    <h3 className="author_comment_title">Jessica Johnson</h3>
-                                    <p className="description_comment">
-                                        Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
-                                    </p>
-                                </div>
-                            }
-                        />
-                        <List
-                            className="comment-list"
-                            itemLayout="horizontal"
-                            dataSource={data}
-                            renderItem={item => (
-                                <li className="item_comment">
-                                    <Rate style={{ color: '#d14031' }} disabled defaultValue={item.rate} />
-                                    <Comment
-                                        className="content_item_comment"
-                                        content={item.content}
-                                    >
-                                    </Comment>
-                                    <div className="bottom_item_comment">
-                                        <img src={item.imageBottomComment} alt="" />
-                                        <span className="right_bottom_item_comment">{item.author}</span>
-                                    </div>
-                                </li>
-                            )}
-                        />
+                        <Comments />
                     </Col>
                 </Row>
                 <Row className="tabs_product_detail">
